requirements.txt
confluent-kafka==2.10.0
requests==2.31.0
python-dotenv==1.0.1
gtfs-realtime-bindings
Geopandas
shapely